==> Notebooks/01.-Exploracion-de-los-Datos-y-Creacion-DataFrame.ipynb <==
# Exploración de los Datos y Creación DataFrame

1. [Exploración de los Datos Disponibles](#exploracion-de-los-datos-disponibles)
<br></br>
1.1 [Exploración de la Estructura de los Archivos](#exploracion-de-la-extructura-de-los-archivos)
<br></br>
2. [Creación del DataFrame Final](#creacion-dataframe-final)
<br></br>
2.1 [Eliminación de Registros con Fecha Superior al '2019-04-01'](#eliminacion-de-registros-con-fecha-superior)
<br></br>
2.2 [Creación del Campo 'peso_final'](#creacion-del-campo-peso_final)
<br></br>
2.3 [Cambio del Nombre de las Variables](#cambio-del-nombre-de-las-variables)
<br></br>
3. [Creación de Archivo con Formato pickle del DataFrame Final](#creacion-del-archivo-con-formato-pickle-del-dataframe-final)
<br></br>
4. [Conclusiones](#conclusiones)

%load_ext watermark
%watermark

## 1. <a id='exploracion-de-los-datos-disponibles'>Exploración de los Datos Disponibles</a>

!ls ../Data/Original-Data/

!ls -sh ../Data/Original-Data/

Se trata de:
* 23 archivos en los que cada uno representa un mes de datos
* Desde 05/2017 hasta 03/2019
* 349MB en total

### 1.1 <a id='exploracion-de-la-extructura-de-los-archivos'>Exploración de la Estructura de los Archivos</a>

import os                           #Funcionalides para interactuar con el sistema de archivos
import pandas as pd                 #Carga de Pandas
import datetime                     #Módulo para el tratamiento de fechas
import numpy as np                  #Carga de Numpy

list_of_files = os.listdir('../Data/Original-Data/')

list_of_files.sort()
list_of_files

Veamos lo que contiene el archivo más antiguo:

!head ../Data/Original-Data/2017.05.csv

¿Tienen todos los archivos una estructura similar?

directory = '../Data/Original-Data/'

for file in list_of_files:
    df_temp = pd.read_csv(directory + file,sep = ";",decimal = ',',encoding = 'utf-8')
    print('El archivo {} tiene la forma {} y contiene {}% NaN\'s'.format(
          file,
          df_temp.shape,
          round(df_temp.isnull().sum().sum() / (df_temp.shape[0] * df_temp.shape[1]),2)
          ))

Trabajamos con el archivo 2017.05.csv. Creamos el primer dataframe y obtenemos información de las variables.

df_temp = pd.read_csv('../Data/Original-Data/2017.05.csv',sep = ";",decimal = ',',encoding = 'utf-8')

df_temp.head()

df_temp.dtypes

print('El archivo 2017.05.csv tiene los siguientes porcentajes de NaN\'s por columnas:\n\
      Startclock Day: {}%\n\
      Product Code: {}%\n\
      Origin Country: {}%\n\
      Destination Country: {}%\n\
      Weight: {}%\n\
      Volumetric Weight: {}%\n\n'.format(round(df_temp['Startclock Day'].isnull().sum() / len(df_temp.index),2),
                                         round(df_temp['Product Code'].isnull().sum() / len(df_temp.index),2),
                                         round(df_temp['Origin Country'].isnull().sum() / len(df_temp.index),2),
                                         round(df_temp['Destination Country'].isnull().sum() / len(df_temp.index),2),
                                         round(df_temp['Weight'].isnull().sum() / len(df_temp.index),2),
                                         round(df_temp['Volumetric Weight'].isnull().sum() / len(df_temp.index),2)
                                         ))

Porcentaje de NaN's de todos los archivos por columnas.

list_of_files = os.listdir('../Data/Original-Data/')
list_of_files.sort()

directory = '../Data/Original-Data/'

for file in list_of_files:
    df_temp = pd.read_csv(directory + file,sep = ";",decimal = ',',encoding = 'utf-8')
    print('El archivo {} tiene los siguientes porcentajes de NaN\'s por columnas:\n\
      Startclock Day: {}%\n\
      Product Code: {}%\n\
      Origin Country: {}%\n\
      Destination Country: {}%\n\
      Weight: {}%\n\
      Volumetric Weight: {}%\n\n'.format(file,
                                         round(df_temp['Startclock Day'].isnull().sum() / len(df_temp.index),2),
                                         round(df_temp['Product Code'].isnull().sum() / len(df_temp.index),2),
                                         round(df_temp['Origin Country'].isnull().sum() / len(df_temp.index),2),
                                         round(df_temp['Destination Country'].isnull().sum() / len(df_temp.index),2),
                                         round(df_temp['Weight'].isnull().sum() / len(df_temp.index),2),
                                         round(df_temp['Volumetric Weight'].isnull().sum() / len(df_temp.index),2)
                                         ))

## 2. <a id='creacion-del-dataframe-final'>Creación del DataFrame Final</a>

Agrupamiento de todos los datos en un único Dataframe. Y sustitución de NaN's:
* Variable 'Startclock Day'. Utilizamos el método de la fecha del registro anterior y convertimos a formato de fecha.
* Variable 'Weight'. Sustituimos por el valor mínimo a facturar (0.5) y convertimos en float.
* Variable 'Volumetric_Weight'. Sustituimos por el valor mínimo a facturar (0.5) y convertimos en float.

df = pd.DataFrame()

list_of_files = os.listdir('../Data/Original-Data/')
list_of_files.sort()

directory = '../Data/Original-Data/'

for file in list_of_files:
    print('Cargando fichero: ',file)
    df_temp = pd.read_csv(directory + file,sep = ";",decimal = ',',encoding = 'utf-8')
    df_temp['Startclock Day'] = df_temp['Startclock Day'].fillna(method = 'ffill')
    df_temp['Startclock Day'] = pd.to_datetime(df_temp['Startclock Day'])
    df_temp['Weight'] = df_temp['Weight'].fillna('0.5')
    df_temp['Weight'] = df_temp['Weight'].astype(float)
    df_temp['Volumetric Weight'] = df_temp['Volumetric Weight'].fillna('0.5')
    df_temp['Volumetric Weight'] = df_temp['Volumetric Weight'].astype(float)
    df = pd.concat([df,df_temp])

df = df.sort_values(by = 'Startclock Day')

df = df.reset_index(drop = True)

df.head()

### 2.1 <a id='eliminacion-de-registros-con-fecha-superior'>Eliminación de Registros con Fecha Superior al '2019-04-01'</a>

df.tail()

Eliminamos envíos con fechas superiores al '2019-04-01', puesto que son envíos recolocados y distorsionan la información que podamos obtener del dataset.

df = df[df['Startclock Day'] <= '2019-04-01']

df.tail()

### 2.2 <a id='creacion-del-campo-peso_final'>Creación del Campo 'peso_final'</a>

Para la facturación del envío se tiene en cuenta el valor de 'Weight' y de 'Volumetric Weight' y se elige el mayor de ambos. Esto es lo que va a determinar el campo 'peso_final'.

df['peso_final'] = np.where(df['Weight'] > df['Volumetric Weight'],df['Weight'],df['Volumetric Weight'])

df.head()

### 2.3 <a id='cambio-del-nombre-de-las-variables'>Cambio del Nombre de las Variables</a>

df.columns = ['Fecha_envio', 
              'Product_code', 
              'Origen',
              'Destino', 
              'Peso', 
              'Volumen',
              'Peso_final']

df.head()

df.shape

df.dtypes

## 3. <a id='creacion-del-archivo-con-formato-pickle-del-dataframe-final'>Creación de Archivo con Formato pickle del DataFrame Final</a>

df.to_pickle('../Data/01.-Exploracion-de-los-Datos-y-Creacion-DataFrame.pkl')

!ls -l --block-size=M ../Data/01.-Exploracion-de-los-Datos-y-Creacion-DataFrame.pkl

## 4. <a id='conclusiones'>Conclusiones</a>

Los archivos de los datos originales tienen las siguientes características:
* 23 archivos en los que cada uno representa un mes de datos
* Desde 05/2017 hasta 03/2019
* 349MB en total

Después de algunas acciones de 'Feature Engineering' el DataFrame final tiene las siguientes características:
* 12486115 de registros y 7 variables
* Ocupa 586MB
* Visualmente su estructura es:

df.head()
