==> Notebooks/02.-Creacion-del-DataFrame-df_U_FR.ipynb <==
# Creación del DataFrame df_U_FR

1. [Carga del DataFrame](#carga-del-dataframe)
<br></br>
2. [Localización del Producto Estrella Teniendo en Cuenta el País](#localizacion-del-producto-estrella-teniendo-en-cuenta-el-pais)
<br></br>
3. [Creación del DataFrame df_U_FR](#creacion-del-dataframe-df_U_FR)
<br></br>
4. [Creación del Archivo con Formato Pickle del DataFrame df_U_FR](#creacion-del-archivo-con-formato-pickle-del-dataframe-df_U_FR)
<br></br>
5. [Conclusiones](#conclusiones)

%load_ext watermark
%watermark

## 1. <a id='carga-del-dataframe'>Carga del DataFrame</a>

import pandas as pd
import datetime
import warnings

warnings.filterwarnings('ignore')

df = pd.read_pickle('../Data/01.-Exploracion-de-los-Datos-y-Creacion-DataFrame-Base.pkl')

df.head()

## 2. <a id='localizacion-del-producto-estrella-teniendo-en-cuenta-el-pais'>Localización del Producto Estrella Teniendo en Cuenta el País</a>

df['Product_code'].unique()

df_shippings = df.groupby(['Destino','Product_code']).size().reset_index()

df_shippings.head()

df_shippings.columns = ['Destino','Producto','Total']

df_shippings = df_shippings.sort_values(by = 'Total',ascending = False)

df_shippings.head()

El 'Producto' N es del que más envíos se realiza, teniendo en cuenta la segmentación de países. Sin embargo es el producto **U para FR** el que vamos a considerar, pues:
* Se realizan prácticamente los mismos envíos que de ES N.
* El margen de ingreso por envío es considerablemente mayor.

df_shippings_FR = df_shippings[df_shippings['Destino'] == 'FR']

df_shippings_FR.head()

df_shippings_FR['Porcentaje_del_total'] = ((df_shippings_FR['Total'] / df_shippings_FR['Total'].sum()) * 100).round(2)

df_shippings_FR.head()

El producto U para Francia representa un 90.13% de los envíos, lo cual refuerza la decisión de considerarlo el producto estrella teniendo en cuenta la segmentación por países.

Vamos a crear un DataFrame únicamente para los campos 'Fecha_envio' y 'Total', para el 'Producto' U en el 'Destino' FR.

## 3. <a id='creacion-del-dataframe-df_U_FR'>Creación del DataFrame df_U_FR</a>

df.head()

df_U_FR = df[(df['Product_code'] == 'U') & (df['Destino'] == 'FR')]

df_U_FR.head()

df_U_FR = df_U_FR.groupby(['Fecha_envio']).size().reset_index()

df_U_FR.head()

Aprovechamos para renombrar las columnas como nos exige Prophet para procesar el dataframe.

df_U_FR.columns = ['ds','y']

df_U_FR.head()

df_U_FR.tail()

## 4. <a id='creacion-del-archivo-con-formato-pickle-del-dataframe-df_U_FR'>Creación del Archivo con Formato Pickle del DataFrame df_U_FR</a>

df_U_FR.to_pickle('../Data/02.-Creacion-del-DataFrame-df_U_FR.pkl')

!ls -l --block-size=M ../Data/02.-Creacion-del-DataFrame-df_U_FR.pkl

## 5. <a id='conclusiones'>Conclusiones</a>

* El producto estrella es U en Francia, que además representa más del 90% de los envíos en este país. Y se ha creado el dataframe que muestra su comportamiento en el rango de fechas 05/2017 hasta 03/2019, para posteriormente procesarlo con Prophet.

* El dataframe df_U_FR ha sido guardado en el archivo '02.-Creacion-del-DataFrame-df_U_FR.pkl'.
