==> Notebooks/04.-Obtencion-Predicciones-por-Producto-y-Pais.ipynb <==
# Obtención Predicciones por Producto y País

0. [Introducción](#introduccion)
<br></br>
1. [Carga del DataFrame](#carga-del-dataframe)
<br></br>
2. [Función que Extrae Datos por País y Producto](#funcion-que-extrae-datos-por-pais-y-producto)
<br></br>
3. [Generación de la Predicción](#generacion-de-la-prediccion)
<br></br>
4. [Conclusiones](#conclusiones)

%load_ext watermark
%watermark

## 0. <a id='introduccion'>Introducción</a>

Siguiendo con la filosofía de este proyecto de **crear predicciones en escala**, se crea este notebook con el que se pretende acelerar todo el proceso para conseguir la predicción.

El objetivo, por tanto, aquí es desarrollar un código que, a partir del DataFrame Base creado en el notebook "01.-Exploracion-de-los-Datos-y-Creacion-DataFrame-Base", se genere automáticamente una predicción con Prophet para un producto ('')

## 1. <a id='carga-del-dataframe'>Carga del DataFrame</a>

import math
import numpy as np
import pandas as pd
from datetime import datetime,timedelta
import calendar
import matplotlib.pyplot as plt
from fbprophet import Prophet
import warnings

%matplotlib inline
plt.rcParams['figure.figsize'] = (18,10)
warnings.filterwarnings('ignore')

df = pd.read_pickle('../Data/01.-Exploracion-de-los-Datos-y-Creacion-DataFrame.pkl')

df.head()

## 2. <a id='funcion-que-extrae-informacion-por-pais-y-producto'>Función que Extrae Datos por País y Producto</a>

def dataframe_creation(product,country):
    df_temp = df[(df['Product_code'] == product) & (df['Destino'] == country)]
    df_temp = df_temp.groupby(['Fecha_envio']).size().reset_index()
    return df_temp

df_N_ES = dataframe_creation('N','ES')

df_N_ES.head()

## 3. <a id='generacion-de-la-prediccion'>Generación de la Predicción</a>

Utilizamos la función creada en el Notebook "03a.-Implementacion-Prophet-a-df_U_FR.ipynb":

def new_y_for_date(date,df_temp):
    df_monthdays_for_date = df_temp.loc[(df_temp['ds'].dt.year == date.year) &
                                        (df_temp['ds'].dt.month == date.month) &
                                        (df_temp['ds'].dt.weekday == date.weekday())]
    new_y = df_monthdays_for_date['y'].mean()
    if math.isnan(new_y) == False:
        return np.int64(round(new_y))
    else:
        return np.int64(round(df_temp.loc[df_temp['ds'].dt.weekday == date.weekday()]['y'].mean()))

Generamos una función a partir de snippets de código del mismo Notebook, que nos genera un dataframe completo y preparado para ser procesado por Prophet:

def dataframe_creation(product,country):
    df_temp = df[(df['Product_code'] == product) & (df['Destino'] == country)]
    df_temp = df_temp.groupby(['Fecha_envio']).size().reset_index()
    df_temp.columns = ['ds','y']
    missing_dates = pd.date_range(start = '2017-05-01',end = '2019-04-01').difference(df_temp['ds'])
    df_missing_dates = pd.DataFrame()
    df_missing_dates['ds'] = missing_dates
    df_missing_dates['y'] = [new_y_for_date(date,df_temp) for date in df_missing_dates['ds']]
    df_temp = df_temp.append(df_missing_dates)
    df_temp['ds'] = pd.to_datetime(df_temp['ds'])
    df_temp = df_temp.sort_values(by = 'ds')
    df_temp = df_temp.reset_index(drop = True)
    return df_temp

Creamos una última función que genera la predicción a partir únicamente del 'Product_code', 'Destino' y el country_name_code, donde éste último ajusta a Prophet para que tenga en cuenta los días de vacaciones del país con el queramos trabajar:

def forecast(product,country,country_name_code):
    f = Prophet()
    f.add_country_holidays(country_name = country_name_code)
    f.fit(dataframe_creation(product,country))
    df_temp = f.make_future_dataframe(periods = 59)
    df_temp = f.predict(df_temp)
    return df_temp

Comprobamos que funciona correctamente:

df_N_ES = forecast('U','FR','FRA')

df_U_FR.head()

## 3. <a id='conclusiones'>Conclusiones</a>
